backend/requirements.txt
# Backend-only dependencies for ReasonSQL API
# Install with: pip install -r backend/requirements.txt

# Core
pydantic>=2.0.0
python-dotenv>=1.0.0

# LLM providers
google-generativeai>=0.3.0
litellm>=1.0.0
groq>=0.4.0

# Agent framework
crewai[tools]>=0.86.0

# API server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0

# Database
psycopg2-binary>=2.9.0

# Utilities
tabulate>=0.9.0
rich>=13.0.0
